# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem manajemen inventory dengan authentication JWT",
        "title": "Inventory Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/barang": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil semua data barang",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barang"
                ],
                "summary": "Get all barang",
                "responses": {
                    "200": {
                        "description": "List semua barang",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Barang"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat data barang baru",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barang"
                ],
                "summary": "Create new barang",
                "parameters": [
                    {
                        "description": "Data barang baru",
                        "name": "barang",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Barang"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Barang berhasil dibuat",
                        "schema": {
                            "$ref": "#/definitions/models.Barang"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/barang/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil data barang berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barang"
                ],
                "summary": "Get barang by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Barang ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data barang",
                        "schema": {
                            "$ref": "#/definitions/models.Barang"
                        }
                    },
                    "400": {
                        "description": "ID tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Barang tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengupdate data barang berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barang"
                ],
                "summary": "Update barang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Barang ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data barang yang akan diupdate",
                        "name": "barang",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Barang"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Barang berhasil diupdate",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus data barang berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Barang"
                ],
                "summary": "Delete barang",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Barang ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Barang berhasil dihapus",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "ID tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Barang": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "kategori_id": {
                    "type": "string"
                },
                "nama": {
                    "type": "string"
                },
                "stok": {
                    "type": "integer"
                },
                "tanggal_buat": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token dengan format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}